[feature/redux] Refactory : Wrap App Component with Category Thunk
