+++ src/store/testPages/TestForRegionCategoryPage.jsx
@@ -1,17 +1,10 @@
 import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import useInitializeRegionCategoryData from "../slices/regionCategoryHook.js";
-import {talentCategoryAction} from "../slices/talentCategorySlice.js";
 import {regionCategoryAction} from "../slices/regionCategorySlice.js";
 
 const TestRegionCategoryPage = () => {
 
     const regionCategoryList = useSelector((state) => state.regionCategory.regionCategories);
-
-    // regionCategory의 초기 데이터는 fetch를 통해 가져와야 하므로,
-    // 처음에 무조건 한 번 fetch를 해 줘야 함
-    // 이 파일은 fetch를 하고 redux에 저장하는 것을 저장해 놓은 hook입니다.
-    useInitializeRegionCategoryData();
 
     const dispatch = useDispatch();
     // dispatch(regionCategoryAction.setCategory('변경된 값'));
